Replacing the main script
